avformat/vorbiscomment: use null buf to calculate vorbis comment length
Also check possible failures when calculating length, and change return type to
int as bigger return values are no longer possible.

+++ libavformat/vorbiscomment.h
@@ -34,8 +34,8 @@
  * For no string, set to an empty string.
  * @return The length in bytes.
  */
-int64_t ff_vorbiscomment_length(const AVDictionary *m, const char *vendor_string,
-                                AVChapter **chapters, unsigned int nb_chapters);
+int ff_vorbiscomment_length(const AVDictionary *m, const char *vendor_string,
+                            AVChapter **chapters, unsigned int nb_chapters);
 
 /**
  * Write a VorbisComment into an AVIOContext. The output size can be obtained
